workflows: Add release in main action
Add the release in the main action.
The release will only be trigger on tags push, only people with
high privileges can push tags, so no security problem.
Having it in a separate action file does not work because the
workflow_run can not know if was a tag push or a pull request.
Having  it together with the main build action improves maintainability
because we do not have to repeat the build step.

Signed-off-by: Eduardo Gonzalez <eduardo.gonzalez@amarulasolutions.com>